+++ src/arduino/lib/BarBot/CConveyor.h
@@ -17,7 +17,7 @@ class CConveyor : public CDispenser
      CConveyor(uint8_t start_stop, uint8_t end_sensor);
      ~CConveyor();
      uint8_t          get_dispener_type();
-     bool             dispense(uint8_t qty);
+     bool             dispense(uint16_t qty);
      bool             loop();
      dispenser_state  get_status();
      void             stop();
