+++ src/arduino/lib/BarBot/CDasher.h
@@ -14,7 +14,7 @@ class CDasher : public CDispenser
      CDasher(uint8_t cam, uint8_t driver);
      ~CDasher();
      uint8_t          get_dispener_type();
-     bool             dispense(uint8_t qty);
+     bool             dispense(uint16_t qty);
      bool             loop();
      dispenser_state  get_status();
      void             stop();
